helper-cli: Move RepositoryPathExcludes.writeAsYaml() to Utils
The function is already used by mutliple commands, so utils is a better
place for it.
