+++ helper-cli/src/main/kotlin/commands/ExportPathExcludesCommand.kt
@@ -26,17 +26,14 @@ import com.github.ajalt.clikt.parameters.options.option
 import com.github.ajalt.clikt.parameters.options.required
 import com.github.ajalt.clikt.parameters.types.file
 
-import java.io.File
-
 import org.ossreviewtoolkit.helper.common.RepositoryPathExcludes
 import org.ossreviewtoolkit.helper.common.getRepositoryPathExcludes
 import org.ossreviewtoolkit.helper.common.mergePathExcludes
 import org.ossreviewtoolkit.helper.common.replaceConfig
+import org.ossreviewtoolkit.helper.common.writeAsYaml
 import org.ossreviewtoolkit.model.OrtResult
 import org.ossreviewtoolkit.model.readValue
-import org.ossreviewtoolkit.model.yamlMapper
 import org.ossreviewtoolkit.utils.expandTilde
-import org.ossreviewtoolkit.utils.safeMkdirs
 
 internal class ExportPathExcludesCommand : CliktCommand(
     help = "Export the path excludes to a path excludes file which maps repository URLs to the path excludes for the " +
@@ -86,18 +83,4 @@ internal class ExportPathExcludesCommand : CliktCommand(
             .mergePathExcludes(localPathExcludes, updateOnlyExisting = updateOnlyExisting)
             .writeAsYaml(pathExcludesFile)
     }
-}
-
-/**
- * Serialize this [RepositoryPathExcludes] to the given [targetFile] as YAML.
- */
-internal fun RepositoryPathExcludes.writeAsYaml(targetFile: File) {
-    targetFile.parentFile.apply { safeMkdirs() }
-
-    yamlMapper.writeValue(
-        targetFile,
-        mapValues { (_, pathExcludes) ->
-            pathExcludes.sortedBy { it.pattern }
-        }.toSortedMap()
-    )
 }
